test(#109): move collection e2e tests to package
